Added config file docs
